requirements.txt
numpy>=1.19.0
torch>=1.7.0
torchvision>=0.8.0
matplotlib>=3.3.0
pyyaml>=5.3.0
scikit-learn>=0.23.0
tqdm>=4.46.0
pandas>=1.0.0
pillow>=7.0.0
tensorboard>=2.3.0